requirements.txt
httpx[http2]>=0.27
feedparser>=6.0
beautifulsoup4>=4.12
readability-lxml>=0.8
lxml>=5.0
python-dotenv>=1.0
